[Ref] rename file with test decorators to mccode_antlr.test
