+++ tests/runtime/test_instr.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 from loguru import logger
-from mccode_antlr.compiled import compiled_test, mcpl_compiled_test
+from mccode_antlr.test import compiled_test, mcpl_compiled_test
 from mccode_antlr.utils import compile_and_run
 from mccode_antlr import Flavor
