+++ tests/runtime/test_when.py
@@ -1,6 +1,6 @@
 from textwrap import dedent
 from mccode_antlr.loader import parse_mcstas_instr
-from mccode_antlr.compiled import compiled_test
+from mccode_antlr.test import compiled_test
 from mccode_antlr.utils import compile_and_run
 
 @compiled_test
